VC Layer Regen setup added
- MDI functions to setup the Regen in VC layer added
- Some missing APIs added

* the ramp is not implemented in the Regen
* Battery reading in PWRT_Init_MC() function does not work

+++ src/GNR2_Firmware/revA/grn2_ra6t2_ftexboard/FTEX/inc/middleware_motor/Regen.h
@@ -48,7 +48,7 @@ bool RegenGetEnabled();
     * @param   hBusVoltage: The bus voltage value.
     * @return  The updated motor speed after applying regen operation.
     */
-uint16_t ApplyRegen(int16_t hMotorSpeed, uint16_t hBusVoltage);
+uint16_t ApplyRegen(int hMotorSpeed, uint16_t hBusVoltage);
 
 /**
     * @brief   Set the regen level as a percentage.
